Add feature to retain texture of a marked area until a task is completed.
Add of list of all assigned areas to the Holding.
SelectionTool is now obtained by a player in multiplayer as well.
Implement Client system for TaskManagement.
Move pushScreen from Authority to Client system.

TODO: Revert back the availableTasks to a Queue instead of List.
TODO: Area selection not working in multiplayer.